fcntl: fix lock type interpretation
The lock type switching from an integer to a string in those
testcases are wrong because of typo I believe. As part of a
batch porting, fcntl17.c did it right.

In /usr/include/asm-generic/fcntl.h:166 defines the posix file
lock type value:

Those testcases mistook them as 1, 2, 3 respectively.

This does not affect the test verdict, but is printing wrong
message when the type is really not match, for example when
testing on CIFS.

Signed-off-by: Murphy Zhou <jencce.kernel@gmail.com>
Reviewed-by: Li Wang <liwang@redhat.com>