sctp_big_chunk: Do not use MD5 when FIPS enabled
MD5 is not FIPS compliant. But still MD5 HMAC algorithm is used as the
default algorithm for sctp even when fips is enabled. Due to this,
sctp_big_chunk testcase is failing because listen() system call in
setup_server() is failing in fips environment.

Fix is to not use MD5 algorithm while setting up server, instead set it
to none.

Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-Off by: Ashwin Dayanand Kamat <kashwindayan@vmware.com>