diotest: Fix the repeated log printing problem
Fix the problem of repeated printing logs when the dio module specifies
the output log file. This is a typical buffer problem. In the
diotest6/diotes3 third step will fork lots of process. If the cache
subprocess is not cleaned up, the cache generated in the previous step
will be copied, and then when each process exits, a large number of
repeated prints will be generated.

Signed-off-by: Duncan.chu <wqhaicyj@163.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>