bpf: use consistent log size
Currently the log size for bpf_prog01 + bpf_prog02 is set to BUFSIZ
which is defined by the libc in stdio.h. In glibc, this is set to the
constant 8192. In other libc's this may be different, e.g. in Musl this
is set to 1024. For bpf_prog02, a log size of 1024 is not large enough
to store the verifier log resulting in the following test breakage:

[...] bpf_common.c:123: TBROK: Failed verification: ENOSPC (28)

This error is returned from kernel/bpf/verifier.c when the verifier log
exceeds the user supplied buffer.

Align bpf_prog01 + bpf_prog02 with other bpf tests and set the buffer
size explicitly to 8192 in bpf_common.h. Since all tests use this value,
use a common definition instead of redefining it in each test.

Signed-off-by: Zachary Leaf <zachary.leaf@arm.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>