sched: add sched sysctl sanity test
This is regression test for sched rr sysctl knobs. The problem with
kernel was that sysctl_sched_rt_period is unsigned int but it was parsed
with proc_dointvec() without explicit bounds which means that you are
allowed to write negative values into the variable even though
documentation says it shouldn't be possible and the kernel code asserted
that rt_period is > 0.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>