fanotify13: Test watching overlayfs upper fs
Run a test variant with overlayfs (over all supported fs)
when watching the upper fs.

This is a regression test for kernel fix bc2473c90fca
("ovl: enable fsnotify events on underlying real files"),
from kernel 6.5, which is not likely to be backported to older kernels.

To avoid waiting for events that won't arrive when testing old kernels,
require that kernel supports encoding fid with new flag AT_HADNLE_FID,
also merged to 6.5 and not likely to be backported to older kernels.

Link: https://lore.kernel.org/ltp/20230903111558.2603332-2-amir73il@gmail.com/
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>
[ pvorel: add git hash, rename variable ]
Signed-off-by: Petr Vorel <pvorel@suse.cz>