lib: tst_test: Fix lockdown detection on missing sysfs
The lockdown detection function reports -1 when secure boot sysfs file
is not present, which is later mistakenly interpreted as secure boot
enabled in tst_test.c.

This causes regression in *_module sycall tests executed on systems when
secureboot is not compiled-in or supported at all.

Check properly if secure boot is enabled by checking that the return
value from these functions is positive.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Li Wang <liwang@redhat.com>
Reviewed-by: Martin Doucha <mdoucha@suse.cz>