memcg: Account for pages in the per-cpu cache
PAGESIZES is one greater than the batch size for charging pages from
the per CPU cache. So with MEM_TO_ALLOC=2*PAGESIZES we have two pages
that are below the threshold for being charged.

Sometimes something triggers a flush and the pages get charged to the
global counter anyway and the test passes. We have seen cases where
the test times out waiting for this to happen.

So this patch sets a lower bound to allow those cases to pass. It'll
probably speed the test up as well.

Signed-off-by: Richard Palethorpe <rpalethorpe@suse.com>
Reviewed-by: Petr Vorel <pvorel@suse.cz>