setsockopt10: Use lapi/{socket,tcp}.h to fix compilation
Distros with glibc-2.26 and older will fail due missing SOL_TLS
in <sys/socket.h> and TCP_ULP in <netinet/tcp.h>. Use fallback
definitions in lapi/socket.h and lapi/tcp.h.

Also replace quotes ("") with sharp brackets (<>) for <netinet/in.h>
(it's a system header, the same was for netinet/tcp.h, but it was
replaced with lapi).

Link: https://lore.kernel.org/ltp/20231020110918.26066-1-pvorel@suse.cz/
Fixes: bdb37aab9 ("Add setsockopt10 TLS ULP UAF CVE-2023-0461")
Reported-by: Martin Doucha <mdoucha@suse.cz>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Martin Doucha <mdoucha@suse.cz>
Signed-off-by: Petr Vorel <pvorel@suse.cz>