Provide a PATH_MAX-long buffer when expecting ENAMETOOLONG
A number of tests check that syscalls manipulating paths return
-ENAMETOOLONG when the specified path is longer than allowed. There
are actually two ways this error can be triggered:

1. If the given string is longer than PATH_MAX, i.e. 4096 as far as
   the kernel is concerned, then the getname() helper will fail and
   the kernel will return -ENAMETOOLONG right away.

2. If the string fits in PATH_MAX, but the filesystem rejects the
   path name, for instance because one of its components is longer
   than the support file name length (e.g. 255 for ext4).

At the moment, we always fail because of 2. since we set PATH_MAX to
1024 in LTP, and in fact we sometimes use NAME_MAX as buffer length
instead (which is shorter). Thus what actually happens is that the
kernel overreads the provided buffer and finds some zero before
hitting the "real" PATH_MAX limit (4096).

We should clearly avoid such overreads. Having the syscalls fail
right away by hitting 1. is the safer option. To do so, we obtain
the value of PATH_MAX directly from the appropriate kernel header,
and ensure it is used where relevant. This affects at least the
following tests:

- rename10
- symlink03

Signed-off-by: Kevin Brodsky <kevin.brodsky@arm.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>