cgroup: Fix scanning V1 mount options
The validation in cgroup_find_ctrl did not take into consideration how
we scan V1 mounts. For V1 we try using each mount option as if it is a
controller name. Some mount options contain characters which would be
invalid in a controller name.

This commit ignores errors when scanning V1 mount options.

Ideally we would have a place where we can read the subsys name
knowing that only subsys names will be read from that location. Such
as with V2. One possibility is /proc/cgroups which contains a numeric
value for the hierarchy. However it would require more investigation.

Fixes: #1093
Fixes: 5292c46e5 "cgroup: Handle trailing new line in cgroup.controllers"
Signed-off-by: Richard Palethorpe <rpalethorpe@suse.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>