open_posix_testsuite: fix pthread_cond_destroy test error.
According to the description of the pthread_cond_destroy function in glibc 2.32,
it is advised to ensure that all threads blocked on the condition variable have
been successfully awakened before calling this function, otherwise unpredictable
situations may occur.

In the glibc 2.32 version, this testcase will hang. Therefore, it is necessary
to ensure that all threads have been successfully awakened before calling
pthread_cond_destroy to destroy the condition variable.

Signed-off-by: tanqinyun.tqy <tanqinyun.tqy@alibaba-inc.com>