syscalls: Remove fork12 test
The fork12 test is meant to be a fork bomb for Linux, but it's old
and it probably made more sense long time ago. Nowadays this test
is mostly killed by the OS, especially when it's running inside
containers. Also, we found that setrlimit01 is doing a better job
for the test scenario, setting a boundary to the maximum number of
processes which can be spawned, then running fork() multiple times
until EAGAIN is reached.

Closes: https://github.com/linux-test-project/ltp/pull/1109
Reviewed-by: Richard Palethorpe <rpalethorpe@suse.com>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.com>