getcwd01: Use syscall directly check invalid argument
getcwd() wrapper in musl behaves differently from glibc and uclibc-ng.
glibc and uclibc-ng malloc() the buffer when passed buffer is NULL.
But musl on NULL buffer ignores the size and uses PATH_MAX.

In order not to depend on exact libc implementation switch to use raw
syscall. This requires to update 2 errno.

Fixes: #1084
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Wei Gao <wegao@suse.com>
[ pvorel: reword commit message, use tst_syscall() ]
Signed-off-by: Petr Vorel <pvorel@suse.cz>