ksm: Add test for smart-scan feature
This adds a new ksm (kernel samepage merging) test to evaluate the new
smart scan feature. It allocates a page and fills it with 'a'
characters. It captures the pages_skipped counter, waits for a few
iterations and captures the pages_skipped counter again. The expectation
is that over 50% of the page scans are skipped (There is only one page
that has KSM enabled and it gets scanned during each iteration and it
cannot be de-duplicated).

Link: https://lore.kernel.org/ltp/20231205180952.3843711-3-shr@devkernel.io/
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Stefan Roesch <shr@devkernel.io>