runtest/ltp-aiodio.part3: Keep only unique records
When ltp-aiodio.part3 was added in f5fb66199, it contained & to run
some of the tests in paralel. It might never worked as expected (it was
removed in 7df75e26a, added again in 01bca9a82 and finally removed in
2a61af975).

But now, after fsx-linux rewrite in 134b08362 there is no
reason to repeat tests (tests in runtests run sequentially),
thus do a final cleanup.

Link: https://lore.kernel.org/ltp/20231215183121.57632-3-pvorel@suse.cz/
Reviewed-by: Li Wang <liwang@redhat.com>
Signed-off-by: Petr Vorel <pvorel@suse.cz>