lapi/io_uring: Fix on older distros
In older distributions the inclusion of linux/fs.h breaks libc
sys/mount.h header. That is because linux/fs.h defines MS_RDONLY to 1
while the libc header uses enum, which means that we end up with the
MS_RDONLY in the libc enum replaced with 1 which ends up in a
compilation failure since after the preprocessor pass the num looks
like:

enum {
	1 = 1,
	...
};

The newer libc actually does #undef for the constants so that this does
not happen anymore, but we have to compatible with older distributions
as well.

Now the only thing that actaully the io_uring header seems to use from
that header is the __kernel_rwf_t which is a typedef to an int and we
have a fallback definition in the lapi/io_uring.h as well. So the safest
and easiest solution is to remove the linux/fs.h include.

Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>