fanotify13: Use generic check for mark type support on filesystem
This test has a specific check for AT_HANDLE_FID file handles, which is
a specific reason for a filesystem (e.g. overlayfs) not supporting
filesystem and mount marks with FAN_REPORT_FID.

We now have a generic check for mark type support on filesystem, so use
the more generic check instead of the specific AT_HANDLE_FID check, do
decide if test case needs to be run.

There are now filesystems other than overlayfs (e.g. fuse) that do not
support filesystem and mount marks with FAN_REPORT_FID, so this test
now runs correctly on those filesystems as well (tested with ntfs-3g).

Reviewed-by: Petr Vorel <pvorel@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>