tst_test_macros.h: Require to pass array size in TST_EXP_FAIL*_ARR()
Although having to pass ARRAY_SIZE() of the expected errnos is not
ideal, it gives more flexibility to the tests allowing to use just
portion of the array (will be used in fanotify14 in the next commit).

It looks to be better than keep introduce yet another functions.

Requires updating 2 tests and 1 library test.

Link: https://lore.kernel.org/ltp/20240326144145.747735-3-pvorel@suse.cz/
Suggested-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Signed-off-by: Petr Vorel <pvorel@suse.cz>