New LTP documentation
The new LTP documentation is meant to use Sphinx as the main
documentation framework, with ReconStructedText as the main language.
All the conversion has been done step-by-step, each chapter at time,
updating English syntax and fixing typos. There are minor improvements,
but overall the structure is the same.
The API chapters is ongoing and it will be completed soon in the future.

The old documentation has been moved into doc/old folder, but it will
removed once new documentation will be completed.

Link: https://lore.kernel.org/ltp/20240404131954.27706-1-andrea.cervesato@suse.de/
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.com>