cputhotplug/doc: Reflect change of cpuhotplug test in doc
The series of cpu_hotplug tests has changed a lot compared to
its first version. However, the part of docs which briefly
describe the test behavior are rarely updated
correspondly. This could confuse people who want to read it
when they use the test for the first time. Update the docs to
make them useful.

Signed-off-by: Yiwei Lin <s921975628@gmail.com>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>