Rewrite msgstress testing suite
msgstress testing suite has been rewritten, taking in consideration
that old code was not working most of the times. The new algorithm
simply generates new SysV IPC messages, spawning a sender and a
receiver which will validate data.

With the new algorithm we also reduce the amount of children which
we spawn, since we only have 2: sender and receiver. This permits to
increase the number of messages to send and to do not be depedent from
system overload.

Signed-off-by: Andrea Cervesato <andrea.cervesato@suse.com>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>
Reviewed-by: Cyril Hrubis <chrubis@suse.cz>