syscalls/msgstress01: Fix the stop logic
The stop flag didn't really work because:

- if queue is full msgsnd() will block in the kernel
- if queue is empty msgrcv() will block in the kernel

And if the other process from the reader-writer pair exits the queue
will be never changed and the test will get stuck and killed by watchdog
timer.

What we need to do is to use IPC_NOWAIT, retry manually (after short
usleep) and handle errors manually as well. In that case no processes
will sleep in kernel and setting the stop flag will actually stop the
test.

Link: https://lore.kernel.org/ltp/20240523133132.13978-2-chrubis@suse.cz/
Acked-by: Andrea Cervesato <andrea.cervesato@suse.com>
Reviewed-by: Martin Doucha <mdoucha@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>