syscalls/msgstress01: Fix timeouts
Make the test exit if runtime has been exhausted before we finished the
requested amount of iterations.

For that to happen we let the main test process to loop while checking
the runtime and set the stop flag if runtime was exhausted. We also need
to separate the stop and fail flag and add counter for finished children.

Also if we exhaust our runtime during the initial fork phase we print a
warning, since we hardly did a meaningful testing in that case.

The changes can be tested with -I parameter, e.g. -I 5 should trigger
the TWARN message and you should be able to get the test to stop in the
message sending phase with larger -I value.

Link: https://lore.kernel.org/ltp/20240523133132.13978-3-chrubis@suse.cz/
Reviewed-by: Martin Doucha <mdoucha@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Cyril Hrubis <chrubis@suse.cz>